Update: - Create app directory to use Next.js 13 app router - Add layout file in app directory - Move homepage from page directory to app directory in the page.tsx file - Move global.scss file to app directory and import it in layout file - Add 'use client' to all component and client side js files
