+++ __tests__/index.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Home from '@/pages/index';
+import Home from '@/pages/home';
 
 describe('Home', () => {
 	it('renders a heading', () => {
