+++ __tests__/snapshot.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import Home from '@/pages/index';
+import Home from '@/pages/home';
 
 it('renders homepage unchanged', () => {
 	// const { container } = render(<Home />)
