+++ components/cards/small-card.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useEffect, useState } from 'react';
 import Link from 'next/link';
 import sc from '../../styles/components/SmallCard.module.scss';
