+++ components/contact-icons.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 
 interface Data {
