+++ components/instructions.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 
 import instructions from '../styles/components/Instructions.module.scss';
