+++ styles/pages/Home.module.scss
@@ -73,9 +73,9 @@
 		transform: translateY(-0px);
 	}
 }
-
+/* 
 ///////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////
+/////////////////////////////////////////////////////////////// */
 
 @media only screen and (min-width: 0px) and (max-width: 479px) and (orientation: portrait) {
 	.homepageDescriptionContainer {
